+++ TidalPy/utilities/graphics/__init__.py
@@ -1,1 +1,3 @@
 from .planet_plot import geotherm_plot
+from .grid_plot import GridPlot
+from .global_map import projection_map